+++ src/cjs/utils/datastructures/DataGrid2dListAdapter.js
@@ -17,19 +17,26 @@ var DataGrid2dListAdapter = /** @class */ (function () {
      * @param xSegmentCount
      * @param ySegmentCount
      */
-    function DataGrid2dListAdapter(list, xSegmentCount, ySegmentCount, minDataValue, maxDataValue, minXValue, maxXValue, minYValue, maxYValue) {
+    function DataGrid2dListAdapter(list, xSegmentCount, ySegmentCount
+    // minDataValue?: T,
+    // maxDataValue?: T,
+    // minXValue?: T,
+    // maxXValue?: T,
+    // minYValue?: T,
+    // maxYValue?: T
+    ) {
         if (xSegmentCount * ySegmentCount > list.length) {
             throw "Cannot create DataGrid2dListAdapter, the list (length " + list.length + ") does not provide enough items for " + xSegmentCount + "x" + ySegmentCount + " required entries.";
         }
         // this.canvas = document.getElementById(canvasId);
         this.xSegmentCount = xSegmentCount;
         this.ySegmentCount = ySegmentCount;
-        this.minDataValue = minDataValue;
-        this.maxDataValue = maxDataValue;
-        this.minXValue = minXValue;
-        this.maxXValue = maxXValue;
-        this.minYValue = minYValue;
-        this.maxYValue = maxYValue;
+        // this.minDataValue = minDataValue;
+        // this.maxDataValue = maxDataValue;
+        // this.minXValue = minXValue;
+        // this.maxXValue = maxXValue;
+        // this.minYValue = minYValue;
+        // this.maxYValue = maxYValue;
         this._dataList = list;
         // @private
         // this._minHeight = 0;